move ruby object to own file to make it more manageable.
