+++ ruby-marshal.cabal
@@ -40,6 +40,7 @@ library
       Data.Ruby.Marshal
     , Data.Ruby.Marshal.Encoding
     , Data.Ruby.Marshal.Get
+    , Data.Ruby.Marshal.RubyObject
     , Data.Ruby.Marshal.Types
     , Data.Ruby.Marshal.Internal.Int
